Create Print the Character
The program must accept a string S followed by a character CH in the next line as the input. The program must print the character as the output. Note: The ENTER key will be stored in the character value. So, include '\n' between string and the character. Boundary Condition(s): 5 <= Length of String S <= 100 
Example Input/Output : 
Input: hello h 
Output: h