Create Sum of Least Significant Bits - M and N
The program must accept two integers M and N as the input. The program must print the sum of the least significant bits in the binary representation of M and N as the output. 
Example Input/Output 1: 
Input: 6 5 
Output: 1 Explanation: The binary representation of 6 is 110 The binary representation of 5 is 101 Here the least significant bits of 6 and 5 are 0 and 1 respectively. The sum of 0 and 1 is 1. Hence the output is 1 
Example Input/Output 2: 
Input: 4 8 
Output: 0