Create Distance Covered
Rahul travelled on a speed of X km/hr. He reached the destination on Y hours. The program must accept the speed X and the time taken Y as the input. The program must print the distance Z (in kms) covered by Rahul as the output.
Example Input/Output: 
Input: 80 5 
Output: 400