Create Difference Between Two Time Periods
The program must accept two time values T1 and T2 in the format of HH:MM as the input. The program must print the difference between T1 and T2 as the output. Note: The hour and the minute of T2 are always greater than the hour and the minute of T1. 
Example Input/Output 1: 
Input: 
9:45 
10:59 
Output: 
1:14 
Example 
Input//Output 2: 
Input: 
1:06 
3:10 
Output: 
2:04