Create Discounted Amount to be Paid
A furniture store offers three different discounts on three different types of furniture. The discount percentage and cost price of all the three types of furniture are passed as input to the program. The program must print the total amount required to buy each one in all types of furniture after the discount with precision up to two decimal places as the output. 
Example Input/Output: 
Input: 
10 500 
20 1500 
17 3400 
Output: 
4472.00