Create Reframe N by Finding Tenth Digit
The program must accept an integer N and a sum value X (sum of unit digit, tenth digit and hundredth digit) as the input. The tenth digit is missing in N. The program must find the tenth digit and print N by reframing it as the output. 
Example Input/Output : 
Input: 1235 12 
Output: 12345