Create Equal sum
The program must accept four integers as the input. The program must print Yes if the sum of the first and the second integer is equal to the sum of the third and the fourth integer. Else the program must print No as the output. 
Example Input/Output 1: 
Input: 2 3 4 1 
Output: Yes 
Example Input/Output 2: 
Input: 10 15 16 8 
Output: No